+++ src/Symfony/Component/Messenger/Attribute/Transport.php
@@ -17,10 +17,7 @@ namespace Symfony\Component\Messenger\Attribute;
 #[\Attribute(\Attribute::TARGET_CLASS | \Attribute::IS_REPEATABLE)]
 class Transport
 {
-    public string $name;
-
-    public function __construct(string $name)
+    public function __construct(public readonly string $name)
     {
-        $this->name = $name;
     }
 }