+++ src/Symfony/Component/Messenger/Tests/Transport/Sender/SendersLocatorTest.php
@@ -24,9 +24,6 @@ use Symfony\Component\Messenger\Transport\Sender\SendersLocator;
 
 class SendersLocatorTest extends TestCase
 {
-    /**
-     * @requires PHP 8
-     */
     public function testAttributeMapping()
     {
         $sender = $this->createMock(SenderInterface::class);